requirements.txt
flask
opencv-python-headless
numpy
face_recognition
requests
supabase
python-dotenv
scipy
cmake
streamlit